refactor(files): split file-tree into folder-node and editable-row modules
